requirements.txt
Flask==3.0.3
Flask-JWT-Extended==4.6.0
Flask-WTF==1.2.1
WTForms==3.1.2
mysql-connector-python==8.4.0
joblib==1.4.2
scikit-learn==1.3.2
numpy==1.26.4
pandas==2.2.2
